requirements.txt
Django==1.6.2
Markdown==2.4
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
South==0.8.4
Twisted-Core==13.2.0
Twisted-Web==13.2.0
Unidecode==0.04.14
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
dj-database-url==0.3.0
django-appconf==0.6
django-classy-tags==0.4
django-cms==3.0
django-email-extras==0.2
django-filter==0.7
django-forms-builder==0.10
django-mptt==0.6.0
django-picklefield==0.3.1
django-polymorphic==0.5.4
django-reversion==1.8.0
django-sekizai==0.7
django-twitter-tag==1.2
djangocms-admin-style==0.2.2
djangocms-text-ckeditor==2.1.4
djangorestframework==2.3.13
duplicity==0.6.23
easy-thumbnails==1.5
html5lib==0.999
httplib2==0.8
lockfile==0.8
lxml==3.3.3
oauthlib==0.6.1
oneconf==0.3.7
pexpect==3.1
piston-mini-client==0.7.5
psycopg2==2.5.3
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pygpgme==0.3
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
python-gnupg==0.3.6
pyxdg==0.25
reportlab==3.0
sessioninstaller==0.0.0
six==1.2.0
software-center-aptd-plugins==0.0.0
sphinx-me==0.2.1
system-service==0.1.6
twitter==1.9.1
unity-lens-photos==1.0
virtualenv==1.11.4
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
